Moving learn_from_seednode up to real middleware.
