+++ nucypher/network/middleware.py
@@ -19,16 +19,41 @@ class RestMiddleware:
             raise RuntimeError("Bad response: {}".format(response.content))
         return response
 
+
+    def learn_from_seednode(self, seednode_metadata, timeout=3, accept_federated_only=False):
+        from nucypher.characters.lawful import Ursula
+        # Pre-fetch certificate
+        self.log.info("Fetching bootnode {} TLS certificate".format(seednode_metadata.checksum_address))
+
+        certificate, filepath = self._get_certificate(seednode_metadata.rest_host, seednode_metadata.rest_port)
+
+        potential_seed_node = Ursula.from_rest_url(self,
+                                                   seednode_metadata.rest_host,
+                                                   seednode_metadata.rest_port,
+                                                   certificate_filepath=filepath,
+                                                   federated_only=True)  # TODO: 466
+
+        if not seednode_metadata.checksum_address == potential_seed_node.checksum_public_address:
+            raise potential_seed_node.SuspiciousActivity(
+                "This seed node has a different wallet address: {} (was hoping for {}).  Are you sure this is a seed node?".format(
+                    potential_seed_node.checksum_public_address,
+                    bootnode.checksum_address))
+        try:
+            potential_seed_node.verify_node(self, accept_federated_only=accept_federated_only)
+        except potential_seed_node.InvalidNode:
+            raise  # TODO: What if our seed node fails verification?
+
+        return potential_seed_node
+
     def _get_certificate(self, hostname, port):
-        bootnode_certificate = ssl.get_server_certificate(hostname, port)
-        certificate = x509.load_pem_x509_certificate(bootnode_certificate.encode(),
+        seednode_certificate = ssl.get_server_certificate(hostname, port)
+        certificate = x509.load_pem_x509_certificate(seednode_certificate.encode(),
                                                      backend=default_backend())
         # Write certificate
-        filename = '{}.{}'.format(bootnode.checksum_address, Encoding.PEM.name.lower())
+        filename = '{}.{}'.format(seednode.checksum_address, Encoding.PEM.name.lower())
         certificate_filepath = os.path.join(self.known_certificates_dir, filename)
         _write_tls_certificate(certificate=certificate, full_filepath=certificate_filepath, force=True)
-        self.log.info("Saved bootnode {} TLS certificate".format(bootnode.checksum_address))
-
+        self.log.info("Saved seednode {} TLS certificate".format(seednode.checksum_address))
 
     def enact_policy(self, ursula, id, payload):
         response = requests.post('https://{}/kFrag/{}'.format(ursula.rest_interface, id.hex()), payload,
